Fix LLP64 and general bugs found during ARM64 Windows porting
Fix several bugs found while porting SBCL to ARM64 Windows.  Most are
related to the LLP64 data model where `long` is 32 bits.

LLP64 type truncation fixes (Windows 64-bit):
- gencgc.c, gencgc-impl.h, pmrgc-impl.h, pmrgc.c, monitor.c:
  gc_card_table_mask declared as `long` but used as a bitmask for
  addresses above 4GB.  Change to sword_t.
- coreparse.c: `1L << nbits` produces a 32-bit result on LLP64.
  Cast to (sword_t)1 for correct 64-bit shift.
- coreparse.c: `sizeof expr/sizeof (type)` has ambiguous parse.
  Use `sizeof(expr)/sizeof(type)` form.
- interrupt.c: `(unsigned long)` cast truncates 64-bit code pointers.
  Use (uintptr_t) instead.

General bug fixes (all platforms):
- immobile-space.c, immobile-space.h: tlsf_page_sso declared as
  `unsigned short*` but page offsets can exceed 65535 on platforms
  with large (>64KB) IMMOBILE_CARD_BYTES.  Change to unsigned int*.
  Also fix malloc/memset size to use sizeof *tlsf_page_sso.
- hopscotch.c: Replace `extern int ffs()` declaration (which may not
  link on all Windows toolchains) with __builtin_ffs inline wrapper.
- room.lisp: (room t) crashes with division-by-zero when a type has
  zero instances.  Guard the per-object size computation with plusp.
- gencgc.c: Guard os_protect call for FIXEDOBJ_SPACE_START against
  zero (fixedobj space disabled on some platforms like ARM64).

+++ src/runtime/immobile-space.h
@@ -43,7 +43,7 @@ text_page_address(low_page_index_t page_num)
 }
 
 extern unsigned char* text_page_genmask;
-extern unsigned short int* tlsf_page_sso;
+extern unsigned int* tlsf_page_sso;
 
 static inline low_page_index_t find_fixedobj_page_index(void *addr)
 {
